ADD: Refactored code. Moved routers to their own file in 'controllers' directory
